Ship no dataset, and give the SASA fixture a molecule
The datasets package promised a Trp-cage trajectory that was never bundled:
TrpCage.traj returned the path to a file that had never existed, from v0.1.0
through v2.4.0. Nothing replaces it -- FastMDXplora fetches a deposited
structure and simulates it, so a reference trajectory in the distribution
would be megabytes of wheel for something one command produces.

A SASA test failed once on one job of fifteen, disagreeing by eight per cent
on a single residue, and could not be reproduced. Two plausible causes are
removed. Its fixture was a cloud of random points with the closest pair 0.074
nm apart -- half a bond, deep inside a van der Waals radius -- where
Shrake-Rupley's inside-or-outside decisions are knife-edge and one flipping
moves an atom's area by a quantum; it is now a peptide built from bond lengths
and angles. And the direct route averaged a float32 array with numpy.mean,
which accumulates in float32, where the grouped route went through pandas in
double: they were never going to agree bit-for-bit, and the direct one lost
digits over a long run. Both are in double.

The assertion now reports both arrays, the rows per residue and the per-frame
values for whichever residues differ, because a bare assert cost a round trip.

+++ src/fastmdxplora/analysis/sasa.py
@@ -133,11 +133,16 @@ class SASA(Analysis):
                 labels = np.array([r.index for r in residues])
             return pd.DataFrame({
                 "residue": labels,
-                "mean_sasa_nm2": sasa.mean(axis=0),
+                # Accumulated in double. `numpy.mean` on a float32 array
+                # sums in float32, so a long run loses digits the per-frame
+                # route does not -- pandas groups in double and the two
+                # answers then differ in their last figures for no reason
+                # anyone reading them could guess.
+                "mean_sasa_nm2": sasa.mean(axis=0, dtype=np.float64),
                 # The spread matters: a residue at 1.0 every frame and one
                 # alternating between 0 and 2 have the same mean and are not
                 # the same thing.
-                "std_sasa_nm2": sasa.std(axis=0),
+                "std_sasa_nm2": sasa.std(axis=0, dtype=np.float64),
             })
 
         # Per-residue: build a long-form table. Residue labels = resSeq
